# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SportMaps BFF API
  version: "1.0.0"
  description: >
    API REST del BFF de SportMaps (Express). Autenticación por Bearer JWT de
    Supabase (header `Authorization: Bearer <token>`). Endpoints de state-change
    sensibles exigen además el header CSRF `X-Requested-With: SportMaps`.
    Los webhooks (Wompi, MercadoPago, WhatsApp, shipping) NO usan JWT: validan
    firma/secreto propio. El endpoint interno de notificaciones valida el header
    `x-notif-secret`.
servers:
  - url: https://sportmaps-bff.onrender.com
    description: Producción
  - url: https://sportmaps-bff-dev.onrender.com
    description: Dev
  - url: http://localhost:3000
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de Supabase Auth del usuario logueado.
    internalSecret:
      type: apiKey
      in: header
      name: x-notif-secret
      description: Secreto compartido para el endpoint interno de despacho (pg_net → BFF).
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/marketplace/checkout/pay:
    post:
      tags: [marketplace]
      summary: Pagar una marketplace_transaction existente
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId]
              properties:
                transactionId: { type: string, format: uuid }
      responses:
        '201': { description: Reference para pagar }
        '400': { description: Datos invalidos }
        '404': { description: Transaccion no encontrada o ya procesada }
        '409': { description: Usuario bloqueado por pagos pendientes }
...
  /api/v1/vendor/profile/payment:
    put:
      tags: [vendor]
      summary: Config de pago del vendedor (onboarding step 2)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_methods: { type: array, items: { type: string } }
                bank_data: { type: object }
      responses:
        '200': { description: Metodos de pago actualizados }
...
  /api/v1/events/{eventId}/delegations/mine/payment:
    post:
      tags: [events]
      summary: Participant declares a payment (proof) for own delegation
      security:
        - bearerAuth: []
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number }
                payment_method: { type: string, default: transfer }
                proof_url: { type: string, nullable: true }
                notes: { type: string, nullable: true }
      responses:
        '201': { description: Created }
        '400': { description: Invalid amount }
        '404': { description: No delegation enrolled }
  /api/v1/events/school-tournaments/{id}/delegations/{delId}/payments/{payId}:
    patch:
      tags: [events]
      summary: Host verifies or rejects a declared payment
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: delId, in: path, required: true, schema: { type: string } }
        - { name: payId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [verify, reject] }
                rejection_reason: { type: string, nullable: true }
      responses:
        '200': { description: OK }
        '400': { description: Invalid action }
        '404': { description: Payment not found }
        '409': { description: Payment already processed }
...
  /api/v1/charge-batches/preview:
    post:
      tags: [cobros-y-pagos]
      summary: Vista previa de una operación (cobros nuevos, pagos de pendientes, descuentos); no escribe
      description: >
        Cuerpo = ChargeBatchRequestSchema (bff/src/services/charge-batches.schemas.ts):
        mode (single|multi), target {kind, ids}, athletes[{type: child|adult|unregistered, id}],
        lines[{category, amount?, due_date, concept, notes?, period?{year,month}, enrollment_id?,
        fee_id?, overage_charge_id?, discount?, exonerate?, pay_amount?, close_mode?}],
        pending[{payment_id, seen{amount, amount_paid}, discount?, waive_late_fee?, exonerate?,
        pay_amount, close_mode}], global_discount?, payment?{method, payment_date, reference?,
        receipt_url?, receipt_sha256?, ocr?}, new_athlete?. Topes: 200 atletas, 600 cobros, 10 líneas.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: 'Resultado de preview_charge_batch (rows_to_create, total_amount, skipped, athletes, warnings_count, preview_hash)' }
        '403': { description: 'Sin permiso en esta escuela (o horas adicionales sin ser dueño)' }
        '404': { description: 'Atleta, grupo, pendiente o excedente de otra escuela' }
        '422': { description: 'Validación' }
  /api/v1/charge-batches:
    post:
      tags: [cobros-y-pagos]
      summary: Confirma la operación en una sola transacción (create_charge_batch); idempotente por client_request_id
      description: >
        Mismo cuerpo que /preview más client_request_id (uuid generado al abrir el modal y
        conservado en reintentos), preview_hash, overrides[{athlete, line_idx, action: force|skip}]
        y notify_families (default false).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: '{ batch_id, duplicated, rows_created, total_amount, payment_ids, paid_ids, partial_ids, adjustments, skipped }' }
        '403': { description: 'Sin permiso' }
        '404': { description: 'Algo de otra escuela' }
        '409': { description: 'PREVIEW_STALE, COBRO_CAMBIO, EN_REVISION, PAGO_EN_CURSO, COBRO_CERRADO, ATLETA_DUPLICADO, PERIODO_DUPLICADO' }
        '422': { description: 'Validación, DESCUENTO_EXCEDE, SOBREPAGO, MULTI_NO_PAGA' }
        '429': { description: 'Más de 10 operaciones en 10 min por usuario o 2.000 cobros por hora por escuela' }
    get:
      tags: [cobros-y-pagos]
      summary: Historial de operaciones («Operaciones»); admin y contador
      security:
        - bearerAuth: []
      parameters:
        - { name: cursor, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
      responses:
        '200': { description: '{ items[], next_cursor }' }
        '403': { description: 'Sin permiso' }
  /api/v1/charge-batches/targets:
    get:
      tags: [cobros-y-pagos]
      summary: Atletas activos de un equipo, categoría o plan (pausados fuera salvo include_paused)
      security:
        - bearerAuth: []
      parameters:
        - { name: kind, in: query, required: true, schema: { type: string, enum: [team, category, plan] } }
        - { name: id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: include_paused, in: query, schema: { type: string, enum: ['true', 'false'] } }
      responses:
        '200': { description: '{ group, athletes[{type, id, name, has_guardian, paused}], count, without_guardian, paused_excluded }' }
        '404': { description: 'Grupo de otra escuela' }
  /api/v1/charge-batches/athlete-search:
    get:
      tags: [cobros-y-pagos]
      summary: Buscador y duplicados antes de «+ Atleta nuevo» (documento o nombre = duplicado; teléfono solo = informativo)
      security:
        - bearerAuth: []
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: doc, in: query, schema: { type: string } }
        - { name: phone, in: query, schema: { type: string } }
      responses:
        '200': { description: '{ items[{table, athlete_type, id, full_name, doc_masked, guardian, matched_by[], es_duplicado, motivo}], has_duplicate }' }
        '422': { description: 'Sin criterio de búsqueda' }
  /api/v1/charge-batches/{id}:
    get:
      tags: [cobros-y-pagos]
      summary: Detalle de una operación con el estado actual de cada cobro y el conteo exacto de lo anulable
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: '{ batch, payments[], annul_preview{annullable_count, annullable_total, kept_count} }' }
        '404': { description: 'No existe en esta escuela' }
  /api/v1/charge-batches/{id}/annul:
    post:
      tags: [cobros-y-pagos]
      summary: Anula lo anulable del lote (pending, overdue, rejected, failed); nunca toca cobros con plata
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, expected_count]
              properties:
                reason: { type: string, minLength: 3, maxLength: 300 }
                expected_count: { type: integer, minimum: 0 }
      responses:
        '200': { description: '{ annulled, kept[{payment_id, status}] }' }
        '409': { description: 'ANNUL_STALE (el conteo cambió) o LOTE_ANULADO' }
...
  /api/v1/payment-adjustments:
    get:
      tags: [cobros-y-pagos]
      summary: Informe de descuentos, condonaciones y exoneraciones (quién, cuánto, motivo, atleta); admin y contador
      security:
        - bearerAuth: []
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: reason, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 500, default: 200 } }
      responses:
        '200': { description: '{ from, to, items[], totals{descuentos, recargo_condonado, exonerado, by_reason}, highlights[] }' }
        '403': { description: 'Sin permiso' }
  /api/v1/payment-adjustments/{id}/revert:
    post:
      tags: [cobros-y-pagos]
      summary: Quitar descuento (revert_payment_adjustment) mientras el cobro siga abierto
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3, maxLength: 300 }
      responses:
        '200': { description: OK }
        '403': { description: 'Sin permiso (el contador no revierte)' }
        '409': { description: 'YA_REVERTIDO, PAGO_EN_CURSO, COBRO_CERRADO' }